Changes to layouts to add more features.
Added the ability to shift the entire generated windows by X amount.
This is useful for setting gaps against monitor edges or for creating
regions to put floating windows.

Also took this opportunity to fix a bug with the main_with_stack layout
and to improve the commenting. It is now much easier to follow how the
layout is created.